Implemento verificación para la eliminación del último valor en DeleteCurrency, refactor menor. Implemento verificación de cantidad de imágenes en ImageSlideManager (Caso: Sin imágenes). Implementación de verificación de último valor disponible y algunos cambios menores en WalletManager. Cambio de ubicación de botones de ayuda. Actualización de texto de ayuda para varias activities.

+++ app/src/main/java/proyectodane/usodeldinero/DeleteCurrencyActivity.java
@@ -9,7 +9,6 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-
 import java.io.File;
 import java.util.ArrayList;
 
@@ -89,6 +88,18 @@ public class DeleteCurrencyActivity extends AppCompatActivity {
      **/
     public void deleteValue (View view) {
 
+        // Verifico si este es el último valor existente, en ese caso, impido su eliminación
+        if(WalletManager.getInstance().isThereOnlyOneValidCurrency(this)){
+
+            // Inhabilito el botón para seleccionar y el botón para eliminar el valor
+            setEnabledFalseSelectAndConfirmButton();
+
+            // Informo sobre la imposibilidad de eliminar y salgo de la activity
+            SnackBarManager sb = new SnackBarManager();
+            sb.showTextIndefiniteOnClickActionStartActivity(findViewById(R.id.coordinatorLayout_deleteCurrency),getString(R.string.error_on_file_delete_last_value),5,MainActivity.class,this);
+            return;
+        }
+
         // Elimino el valor del registro y también los posibles valores de este mismo tipo que pudieran existir en la billetera
         WalletManager.getInstance().deleteExistingCurrency(this,st_valueID);
 
@@ -97,11 +108,8 @@ public class DeleteCurrencyActivity extends AppCompatActivity {
         File file = new File(dir,st_valueID);
         boolean deleteOK = file.delete();
 
-        // Inhabilito botón para seleccionar y eliminar el valor
-        Button selectButton = (Button) findViewById(R.id.button25);
-        selectButton.setEnabled(false);
-        Button confirmButton = (Button) findViewById(R.id.button26);
-        confirmButton.setEnabled(false);
+        // Inhabilito el botón para seleccionar y el botón para eliminar el valor
+        setEnabledFalseSelectAndConfirmButton();
 
         // Informo sobre error, en caso de existir
         if(!deleteOK){
@@ -122,6 +130,16 @@ public class DeleteCurrencyActivity extends AppCompatActivity {
     public void showHelp(View view) {
         SnackBarManager sb = new SnackBarManager();
         sb.showTextIndefiniteOnClickActionDisabled(findViewById(R.id.coordinatorLayout_deleteCurrency),getString(R.string.help_text_delete_currency),10);
+    }
+
+    /**
+     * Inhabilito el botón para seleccionar y el botón para eliminar el valor
+     **/
+    private void setEnabledFalseSelectAndConfirmButton(){
+        Button selectButton = (Button) findViewById(R.id.button25);
+        selectButton.setEnabled(false);
+        Button confirmButton = (Button) findViewById(R.id.button26);
+        confirmButton.setEnabled(false);
     }
 
 }
